chore: move scripts from root directory to api directory
