+++ api/package.json
@@ -47,6 +47,7 @@
     "@1024pix/oppsy": "^1.0.1",
     "@adminjs/hapi": "^7.0.1",
     "@adminjs/sequelize": "^4.0.0",
+    "@fast-csv/parse": "^5.0.5",
     "@googleapis/sheets": "^12.0.0",
     "@hapi/boom": "^10.0.0",
     "@hapi/cookie": "^12.0.0",
@@ -76,6 +77,7 @@
     "phrase-js": "^3.0.0",
     "pino": "^10.0.0",
     "pino-pretty": "^13.0.0",
+    "progress": "^2.0.3",
     "qs": "^6.10.1",
     "sequelize": "^6.29.0",
     "showdown": "^2.0.0",
