+++ api/scripts/populate-alpha-and-delta-column-with-csv.js
