+++ api/scripts/reverse-static-courses-challenges-order.js
