--- scripts/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "migrate-attachments-from-airtable",
-  "version": "1.215.2",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "NODE_ENV=test vitest run",
-    "test:watch": "NODE_ENV=test vitest",
-    "lint": "eslint ."
-  },
-  "author": "",
-  "license": "AGPL-3.0-or-later",
-  "dependencies": {
-    "@fast-csv/parse": "^5.0.0",
-    "airtable": "^0.12.1",
-    "lodash": "^4.17.21",
-    "pg": "^8.9.0",
-    "progress": "^2.0.3"
-  },
-  "devDependencies": {
-    "@eslint/js": "^9.38.0",
-    "@stylistic/eslint-plugin": "^5.5.0",
-    "@types/pg": "^8.6.6",
-    "eslint": "^9.38.0",
-    "globals": "^16.0.0",
-    "vitest": "^4.0.0"
-  }
-}
